+++ include/vtzero/point.hpp
@@ -52,16 +52,16 @@ namespace vtzero {
         }
 
         /// Get z value (dummy function for 2d coordinates)
-        int32_t get_z() const noexcept {
+        int64_t get_z() const noexcept {
             return 0;
         }
 
         /// Add value to z value (dummy function for 2d coordinates)
-        void add_to_z(int32_t /*value*/) const noexcept {
+        void add_to_z(int64_t /*value*/) const noexcept {
         }
 
         /// Set z value (dummy function for 2d coordinates)
-        void set_z(int32_t /*value*/) const noexcept {
+        void set_z(int64_t /*value*/) const noexcept {
         }
 
     }; // struct point
@@ -82,30 +82,30 @@ namespace vtzero {
         int32_t y = 0;
 
         /// elevation
-        int32_t z = 0;
+        int64_t z = 0;
 
         /// Default construct to 0 coordinates
         constexpr point<3>() noexcept = default;
 
         /// Constructor
-        constexpr point<3>(int32_t x_, int32_t y_, int32_t z_ = 0) noexcept :
+        constexpr point<3>(int32_t x_, int32_t y_, int64_t z_ = 0) noexcept :
             x(x_),
             y(y_),
             z(z_) {
         }
 
         /// Get z value
-        int32_t get_z() const noexcept {
+        int64_t get_z() const noexcept {
             return z;
         }
 
         /// Add value to z value
-        void add_to_z(int32_t value) noexcept {
-            z = static_cast<int32_t>(static_cast<int64_t>(z) + static_cast<int64_t>(value));
+        void add_to_z(int64_t value) noexcept {
+            z = z + value;
         }
 
         /// Set z value
-        void set_z(int32_t value) noexcept {
+        void set_z(int64_t value) noexcept {
             z = value;
         }
